+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- **Breaking**: 最小サポートをbeta44に引き上げ
 - **Breaking**: `aviutl2::filter::FilterPluginFlags`の`as_object`を`input`、`as_filter`を`filter`に改名
 - generic: `ObjectLayerFrame::frame_range`、`ObjectLayerFrame::frame_range_inclusive`を追加
 - generic: `EditSectionLayersIterator`に`DoubleEndedIterator`を実装